fix(ioredis): remove dependency on ioredis from instrumentation
