+++ plugins/node/opentelemetry-instrumentation-ioredis/package.json
@@ -59,6 +59,7 @@
     "@types/sinon": "10.0.9",
     "@types/node": "18.11.7",
     "cross-env": "7.0.3",
+    "ioredis": "5.2.2",
     "gts": "3.1.0",
     "mocha": "7.2.0",
     "nyc": "15.1.0",
@@ -72,8 +73,7 @@
     "@opentelemetry/instrumentation": "^0.35.1",
     "@opentelemetry/redis-common": "^0.34.0",
     "@opentelemetry/semantic-conventions": "^1.0.0",
-    "@types/ioredis4": "npm:@types/ioredis@^4.28.10",
-    "ioredis": "5.2.2"
+    "@types/ioredis4": "npm:@types/ioredis@^4.28.10"
   },
   "homepage": "https://github.com/open-telemetry/opentelemetry-js-contrib/tree/main/plugins/node/opentelemetry-instrumentation-ioredis#readme"
 }
